+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 ## [Unreleased]
 
-### Breaking Changes
+### Added
+
+- Add support for Elliptic Curve Digital Signature Algorithm ECDSA SHA256
+- Add support for x509 certificate
+- Add possibility to sign specific parts of an XML document
+
+### Changed
 
 - Require PHP 8